+++ CHANGELOG.md
@@ -2,6 +2,7 @@ Changelog
 =========
 
 ## 0.0.2 (2019-07-22)
+ * Added jQuery Terminal assets
  * Added assets and configure grunt
 
 ## 0.0.1 (2019-07-21)